requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
python-multipart==0.0.22
transformers==5.0.0
torch
pillow==11.3.0
omegaconf==2.3.0
requests
nest-asyncio==1.6.0